Update rotary embedding as common function
