+++ optimum/habana/transformers/modeling_utils.py
@@ -45,7 +45,6 @@ from .models import (
     GaudiFalconForCausalLM,
     GaudiFalconMLP,
     GaudiFalconModel,
-    GaudiFalconRotaryEmbedding,
     GaudiGemmaAttention,
     GaudiGemmaDecoderLayer,
     GaudiGemmaForCausalLM,
@@ -64,7 +63,6 @@ from .models import (
     GaudiGPTNeoXAttention,
     GaudiGPTNeoXForCausalLM,
     GaudiGPTNeoXLayer,
-    GaudiGPTNeoXRotaryEmbedding,
     GaudiLlamaAttention,
     GaudiLlamaDecoderLayer,
     GaudiLlamaDynamicNTKScalingRotaryEmbedding,
@@ -89,6 +87,8 @@ from .models import (
     GaudiMptModel,
     GaudiOPTForCausalLM,
     GaudiOPTLearnedPositionalEmbedding,
+    GaudiPersimmonAttention,
+    GaudiPersimmonDecoderLayer,
     GaudiPersimmonForCausalLM,
     GaudiPhiAttention,
     GaudiPhiDecoderLayer,
@@ -99,6 +99,7 @@ from .models import (
     GaudiQwen2ForCausalLM,
     GaudiQwen2MLP,
     GaudiQwen2Model,
+    GaudiStableLmAttention,
     GaudiStableLmDecoderLayer,
     GaudiStableLmForCausalLM,
     GaudiStarcoder2Attention,
@@ -166,8 +167,6 @@ from .models import (
     gaudi_opt_decoder_layer_forward,
     gaudi_opt_model_forward,
     gaudi_owlvitclasspredictionhead_forward,
-    gaudi_persimmon_attention_forward,
-    gaudi_persimmon_decoder_layer_forward,
     gaudi_persimmon_model_forward,
     gaudi_qwen2_rmsnorm_forward,
     gaudi_rot_matmul,
@@ -185,7 +184,6 @@ from .models import (
     gaudi_SpeechT5Attention_forward,
     gaudi_SpeechT5Decoder_forward,
     gaudi_SpeechT5DecoderLayer_forward,
-    gaudi_stablelm_attention_forward,
     gaudi_stablelm_model_forward,
     gaudi_t5_layernorm_forward,
     gaudi_T5Attention_forward,
@@ -371,7 +369,6 @@ def adapt_transformers_to_gaudi():
     transformers.models.gpt_neox.modeling_gpt_neox.GPTNeoXAttention = GaudiGPTNeoXAttention
     transformers.models.gpt_neox.modeling_gpt_neox.GPTNeoXForCausalLM = GaudiGPTNeoXForCausalLM
     transformers.models.gpt_neox.modeling_gpt_neox.GPTNeoXLayer = GaudiGPTNeoXLayer
-    transformers.models.gpt_neox.modeling_gpt_neox.GPTNeoXRotaryEmbedding = GaudiGPTNeoXRotaryEmbedding
     transformers.models.gpt_neox.modeling_gpt_neox.GPTNeoXModel.forward = gaudi_gpt_neox_model_forward
 
     # Optimization for llama generation on Gaudi
@@ -408,7 +405,6 @@ def adapt_transformers_to_gaudi():
     transformers.models.falcon.modeling_falcon.FalconMLP = GaudiFalconMLP
     transformers.models.falcon.modeling_falcon.FalconModel = GaudiFalconModel
     transformers.models.falcon.modeling_falcon.FalconDecoderLayer = GaudiFalconDecoderLayer
-    transformers.models.falcon.modeling_falcon.FalconRotaryEmbedding = GaudiFalconRotaryEmbedding
     transformers.models.falcon.modeling_falcon.FalconLinear.forward = gaudi_falcon_linear_forward
 
     # Optimization for t5 on Gaudi
@@ -483,12 +479,10 @@ def adapt_transformers_to_gaudi():
     transformers.models.speecht5.modeling_speecht5._generate_speech = gaudi_generate_speech
 
     # Optimization for persimmon on Gaudi
+    transformers.models.persimmon.modeling_persimmon.PersimmonAttention = GaudiPersimmonAttention
+    transformers.models.persimmon.modeling_persimmon.PersimmonDecoderLayer = GaudiPersimmonDecoderLayer
     transformers.models.persimmon.modeling_persimmon.PersimmonForCausalLM = GaudiPersimmonForCausalLM
     transformers.models.persimmon.modeling_persimmon.PersimmonModel.forward = gaudi_persimmon_model_forward
-    transformers.models.persimmon.modeling_persimmon.PersimmonAttention.forward = gaudi_persimmon_attention_forward
-    transformers.models.persimmon.modeling_persimmon.PersimmonDecoderLayer.forward = (
-        gaudi_persimmon_decoder_layer_forward
-    )
 
     # Optimization for seamless m4t on Gaudi
     transformers.models.seamless_m4t.modeling_seamless_m4t.SeamlessM4TAttention.forward = (
@@ -544,10 +538,10 @@ def adapt_transformers_to_gaudi():
     transformers.models.qwen2.modeling_qwen2.Qwen2RMSNorm.forward = gaudi_qwen2_rmsnorm_forward
 
     # Optimization for stablelm on Gaudi
+    transformers.models.stablelm.modeling_stablelm.StableLmAttention = GaudiStableLmAttention
+    transformers.models.stablelm.modeling_stablelm.StableLmDecoderLayer = GaudiStableLmDecoderLayer
     transformers.models.stablelm.modeling_stablelm.StableLmForCausalLM = GaudiStableLmForCausalLM
     transformers.models.stablelm.modeling_stablelm.StableLmModel.forward = gaudi_stablelm_model_forward
-    transformers.models.stablelm.modeling_stablelm.StableLmAttention.forward = gaudi_stablelm_attention_forward
-    transformers.models.stablelm.modeling_stablelm.StableLmDecoderLayer = GaudiStableLmDecoderLayer
 
     transformers.models.vision_encoder_decoder.modeling_vision_encoder_decoder.VisionEncoderDecoderModel.prepare_inputs_for_generation = gaudi_VisionEncoderDecoderModel_prepare_inputs_for_generation
