+++ optimum/habana/transformers/models/falcon/__init__.py
@@ -4,6 +4,5 @@ from .modeling_falcon import (
     GaudiFalconForCausalLM,
     GaudiFalconMLP,
     GaudiFalconModel,
-    GaudiFalconRotaryEmbedding,
     gaudi_falcon_linear_forward,
 )
