+++ optimum/habana/transformers/models/gpt_neox/__init__.py
@@ -2,6 +2,5 @@ from .modeling_gpt_neox import (
     GaudiGPTNeoXAttention,
     GaudiGPTNeoXForCausalLM,
     GaudiGPTNeoXLayer,
-    GaudiGPTNeoXRotaryEmbedding,
     gaudi_gpt_neox_model_forward,
 )
