+++ optimum/habana/transformers/models/qwen2/modeling_qwen2.py
@@ -37,6 +37,8 @@ from transformers.models.qwen2.modeling_qwen2 import (
 )
 from transformers.utils import is_torchdynamo_compiling
 
+from optimum.habana.transformers.generation.utils import GaudiRotaryEmbedding
+
 from ...modeling_attn_mask_utils import (
     _gaudi_prepare_4d_causal_attention_mask,
 )
@@ -195,6 +197,7 @@ class GaudiQwen2Attention(Qwen2Attention):
         self.inp_seq_len = -1
         self.norm_factor = 1.0 / math.sqrt(self.head_dim)
         self.block_size = 4096
+        self.rotary_emb = GaudiRotaryEmbedding(config=self.config)
 
     def allocate_kv_cache(self, batch_size, max_seq_len, inp_seq_len):
         cache_shape = (batch_size, self.num_key_value_heads, max_seq_len, self.head_dim)
@@ -950,4 +953,4 @@ def apply_customized_rope(q, k, cos, sin, position_ids):
             k, cos.unsqueeze(0).unsqueeze(0).clone(), sin.unsqueeze(0).unsqueeze(0).clone(), position_ids
         )
     else:
-        return apply_rotary_pos_emb(q, k, cos, sin, position_ids)
+        return apply_rotary_pos_emb(q, k, cos[position_ids], sin[position_ids])
