+++ optimum/habana/transformers/models/stablelm/__init__.py
@@ -1,6 +1,6 @@
 from .modeling_stablelm import (
+    GaudiStableLmAttention,
     GaudiStableLmDecoderLayer,
     GaudiStableLmForCausalLM,
-    gaudi_stablelm_attention_forward,
     gaudi_stablelm_model_forward,
 )
